+++ tests/test_parse_general.py
@@ -74,6 +74,19 @@ def test_parse_module() -> None:
     assert_that(result).is_equal_to(expected_result)
 
 
+def test_parse_type() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "basic/06-type"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BICEP_PARSER.parse(file_path=file_path)
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
 def test_parse_param_decorator() -> None:
     # given
     sub_dir_path = EXAMPLES_DIR / "complex/01-param-decorator"