+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `loadYamlContent` function
 - Support `flatten` function
 
+# Changed
+
+- Fixed multiline issue with union function
+
 # Removed
 
 - Drop support of Python 3.7