+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Added
 
+- Support Python 3.13 officially
 - Support `loadYamlContent` function
 - Support `flatten` function
 - Initial support of custom data types