+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `loadYamlContent` function
 - Support `flatten` function
 - Initial support of custom data types
+- Support safe-dereference operator
 
 # Changed